+++ src/main/java/scanner/XBruteRunner.java
@@ -4,14 +4,14 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import scanner.analyze.ProblemResolver;
 import scanner.brute.BruteForceScanner;
-import scanner.stat.TimeStatGatherer;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-import static scanner.stat.TimeStatItem.*;
+import static scanner.stat.StatDataHolder.TIME_GATHERER;
+import static scanner.stat.TimeStatItem.TOTAL_BRUTE_TIME;
 
 /**
  * Find password camera logic class.
@@ -42,7 +42,7 @@ public class XBruteRunner extends AbstractRunner {
         } catch (Exception xep) {
             log.error("Error during brute attack: {}", xep.getMessage());
         }
-        TimeStatGatherer.set(TOTAL_BRUTE_TIME, timer.elapsed(TimeUnit.MILLISECONDS));
+        TIME_GATHERER.set(TOTAL_BRUTE_TIME, (double) timer.elapsed(TimeUnit.MILLISECONDS));
         timer.stop();
     }