+++ src/main/java/scanner/stat/EfficiencyItem.java
@@ -2,8 +2,8 @@ package scanner.stat;
 
 public enum EfficiencyItem {
 
-    SCAN_EFFICIENCY,
-    BRUTE_EFFICIENCY,
-    GET_FRAME_EFFICIENCY;
+    SCAN_EFFICIENCY { @Override public String toString() { return "Scanning efficiency"; } },
+    BRUTE_EFFICIENCY { @Override public String toString() { return "Brute efficiency"; } },
+    GET_FRAME_EFFICIENCY { @Override public String toString() { return "Getting frame efficiency"; } },
 
 }