+++ src/main/java/scanner/stat/ScreenStatItem.java
@@ -12,20 +12,17 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public enum ScreenStatItem {
 
-    ALL(0),
-    SUCCESS(1),
-    FAILURE(1),
-
-    INVALID_DATA_FOUND(2),
-    WRONG_AUTH_ERROR(2),
-    NOT_FOUND_CODEC_ERROR(2),
-    STREAM_NOT_FOUND(2),
-    BAD_REQUEST(2),
-    PROTOCOL_NOT_SUPPORTED(2),
-    PARAMETER_NOT_UNDERSTOOD(2),
-    UNEXPECTED_ERROR(2),
-    OTHER(2);
-
-    private final int order;
+    ALL { @Override public String toString() { return "All"; } },
+    SUCCESS { @Override public String toString() { return "Success"; } },
+    FAILURE { @Override public String toString() { return "Failure"; } },
+    INVALID_DATA_FOUND { @Override public String toString() { return "Invalid data found"; } },
+    WRONG_AUTH_ERROR { @Override public String toString() { return "Wrong auth error"; } },
+    NOT_FOUND_CODEC_ERROR { @Override public String toString() { return "Not found codec error"; } },
+    STREAM_NOT_FOUND { @Override public String toString() { return "Stream not found"; } },
+    BAD_REQUEST { @Override public String toString() { return "Bad request"; } },
+    PROTOCOL_NOT_SUPPORTED { @Override public String toString() { return "Protocol not supported"; } },
+    PARAMETER_NOT_UNDERSTOOD { @Override public String toString() { return "Parameter not understood"; } },
+    UNEXPECTED_ERROR { @Override public String toString() { return "Unexpected error"; } },
+    OTHER { @Override public String toString() { return "Other"; } }
 
 }